requirements.txt
fastapi
uvicorn[standard]
requests
google-generativeai
python-dotenv
pydantic